refactor: extract golden gub and feedback modules
